Monitor dev

+++ src/main/java/com/gate/gatelib/repository/GroupDao.java
@@ -3,6 +3,6 @@ package com.gate.gatelib.repository;
 import com.gate.gatelib.models.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupDao extends JpaRepository<Group, Integer> {
+public interface GroupDao extends JpaRepository<Group, Long> {
 
 }
